backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2

djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

django-cors-headers==4.4.0

django-filter==24.3

python-decouple==3.8

psycopg2-binary==2.9.9

redis==5.0.8
django-redis==5.4.0

channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

openai==1.51.2
better-profanity==0.7.0

google-auth==2.35.0
requests==2.32.3

Pillow==10.4.0

drf-spectacular==0.27.2

sentry-sdk==2.14.0

pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1
